use switch bind action component
